HDFS-3617. Port HDFS-96 to branch-1 (support blocks greater than 2GB). Contributed by Patrick Kling and Harsh J.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/common/branches/branch-1.1@1385343 13f79535-47bb-0310-9956-ffa450edef68

+++ src/hdfs/org/apache/hadoop/hdfs/DFSClient.java
@@ -2396,7 +2396,7 @@ public class DFSClient implements FSConstants, java.io.Closeable {
             if (pos > blockEnd) {
               currentNode = blockSeekTo(pos);
             }
-            int realLen = Math.min(len, (int) (blockEnd - pos + 1));
+            int realLen = (int) Math.min((long) len, (blockEnd - pos + 1L));
             int result = readBuffer(buf, off, realLen);
             
             if (result >= 0) {
